# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: JieDanBa OPC Matching Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/opc-leaderboard:
    get:
      operationId: getOpcLeaderboard
      tags: [users]
      summary: Get OPC leaderboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpcProfile"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}/opc-profile:
    get:
      operationId: getOpcProfile
      tags: [users]
      summary: Get OPC profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OPC profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpcProfile"
    put:
      operationId: updateOpcProfile
      tags: [users]
      summary: Update OPC profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOpcProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpcProfile"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [in_progress, pending_acceptance, completed, closed, disputed]
        - name: role
          in: query
          schema:
            type: string
            enum: [opc, publisher]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /orders/{orderId}:
    get:
      operationId: getOrderById
      tags: [orders]
      summary: Get order detail
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders/{orderId}/accept:
    post:
      operationId: acceptOrder
      tags: [orders]
      summary: Accept order delivery
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: number
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
      responses:
        "200":
          description: Order accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/reject:
    post:
      operationId: rejectDelivery
      tags: [orders]
      summary: Reject delivery with feedback
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required:
                - reason
      responses:
        "200":
          description: Delivery rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /portfolios:
    get:
      operationId: listPortfolios
      tags: [portfolios]
      summary: List portfolio items
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Portfolio items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Portfolio"
    post:
      operationId: createPortfolio
      tags: [portfolios]
      summary: Add a portfolio item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioInput"
      responses:
        "201":
          description: Portfolio created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
  /portfolios/{portfolioId}:
    put:
      operationId: updatePortfolio
      tags: [portfolios]
      summary: Update a portfolio item
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioInput"
      responses:
        "200":
          description: Updated portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
    delete:
      operationId: deletePortfolio
      tags: [portfolios]
      summary: Delete a portfolio item
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        nickname:
          type: string
        phone:
          type: string
        avatar:
          type: string
        role:
          type: string
          enum: [opc, publisher, admin]
        status:
          type: string
          enum: [active, suspended, banned]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nickname
        - role
        - status
        - createdAt
    OpcProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        nickname:
          type: string
        avatar:
          type: string
        level:
          type: string
          enum: [C, B, A]
        bio:
          type: string
        skillTags:
          type: array
          items:
            type: string
        industryTags:
          type: array
          items:
            type: string
        creditScore:
          type: number
        totalOrders:
          type: integer
        completionRate:
          type: number
        avgRating:
          type: number
        totalEarnings:
          type: number
        activityScore:
          type: number
      required:
        - id
        - userId
        - nickname
        - level
        - creditScore
    UpdateOpcProfileInput:
      type: object
      properties:
        bio:
          type: string
        skillTags:
          type: array
          items:
            type: string
        industryTags:
          type: array
          items:
            type: string
    Milestone:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        deadline:
          type: string
          format: date
        deliverableDesc:
          type: string
        status:
          type: string
          enum: [pending, submitted, approved, rejected]
      required:
        - name
        - deadline
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNo:
          type: string
        demandId:
          type: integer
        demandTitle:
          type: string
        demandType:
          type: string
        opcId:
          type: integer
        opcNickname:
          type: string
        publisherId:
          type: integer
        publisherName:
          type: string
        amount:
          type: number
        opcShare:
          type: number
        publisherShare:
          type: number
        platformFee:
          type: number
        status:
          type: string
          enum: [in_progress, pending_acceptance, completed, closed, disputed]
        deliverables:
          type: array
          items:
            $ref: "#/components/schemas/Deliverable"
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
        rating:
          type: number
        reviewComment:
          type: string
        deadline:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orderNo
        - demandId
        - demandTitle
        - opcId
        - publisherId
        - amount
        - status
        - createdAt
    OrderListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - items
        - total
        - page
        - limit
        - totalPages
    Deliverable:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        milestoneId:
          type: integer
        title:
          type: string
        description:
          type: string
        fileUrl:
          type: string
        fileName:
          type: string
        status:
          type: string
          enum: [submitted, approved, rejected]
        feedback:
          type: string
        submittedAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - title
        - status
        - submittedAt
...
    Portfolio:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        type:
          type: string
        coverImage:
          type: string
        description:
          type: string
        projectUrl:
          type: string
        orderId:
          type: integer
        rating:
          type: number
        clientFeedback:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - type
        - description
        - createdAt
    CreatePortfolioInput:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
        coverImage:
          type: string
        description:
          type: string
        projectUrl:
          type: string
      required:
        - title
        - type
        - description
...