# lib/api-zod/src/generated/types/deliverable.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JieDanBa OPC Matching Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { DeliverableStatus } from "./deliverableStatus";

export interface Deliverable {
  id: number;
  orderId: number;
  milestoneId?: number;
  title: string;
  description?: string;
  fileUrl?: string;
  fileName?: string;
  status: DeliverableStatus;
  feedback?: string;
  submittedAt: Date;
}