# lib/api-zod/src/generated/types/opcProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JieDanBa OPC Matching Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { OpcProfileLevel } from "./opcProfileLevel";

export interface OpcProfile {
  id: number;
  userId: number;
  nickname: string;
  avatar?: string;
  level: OpcProfileLevel;
  bio?: string;
  skillTags?: string[];
  industryTags?: string[];
  creditScore: number;
  totalOrders?: number;
  completionRate?: number;
  avgRating?: number;
  totalEarnings?: number;
  activityScore?: number;
}