# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JieDanBa OPC Matching Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { Deliverable } from "./deliverable";
import type { Milestone } from "./milestone";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  orderNo: string;
  demandId: number;
  demandTitle: string;
  demandType?: string;
  opcId: number;
  opcNickname?: string;
  publisherId: number;
  publisherName?: string;
  amount: number;
  opcShare?: number;
  publisherShare?: number;
  platformFee?: number;
  status: OrderStatus;
  deliverables?: Deliverable[];
  milestones?: Milestone[];
  rating?: number;
  reviewComment?: string;
  deadline?: Date;
  createdAt: Date;
  updatedAt?: Date;
}