# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JieDanBa OPC Matching Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { UserRole } from "./userRole";
import type { UserStatus } from "./userStatus";

export interface User {
  id: number;
  nickname: string;
  phone?: string;
  avatar?: string;
  role: UserRole;
  status: UserStatus;
  createdAt: Date;
}